minor database modifications

+++ database/migrations/2023_04_25_183640_create_fixtures_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('fixtures', function (Blueprint $table) {
             $table->id();
-            $table->integer('fixture_id', false, true)->unique();
+            $table->unsignedSmallInteger('fixture_id', false, true)->unique();
             $table->string('home');
             $table->string('away');
             $table->string('league');
